protected routes

+++ src/controllers/contact.controller.ts
@@ -6,11 +6,11 @@ const mapContact = (contact: ContactDoc) => ({
     id: contact.id,
     name: contact.name,
     email: contact.email,
-    phone: contact.phone
+    phone: contact.phone,
 });
 
-export const getContacts = async () => {
-    const contacts = await Contact.find();
+export const getContacts = async (userId: string) => {
+    const contacts = await Contact.find({userId: userId});
     if(!contacts){
         throw new BadRequest('Unable to get contacts');
     }
@@ -18,7 +18,7 @@ export const getContacts = async () => {
     return contacts.map(mapContact);
 }
 
-export const createContact = async (params: CreateContactDto) => {
+export const createContact = async (userId: string, params: CreateContactDto) => {
 
     const foundContact = await Contact.findOne({
         phone: params.phone
@@ -27,8 +27,8 @@ export const createContact = async (params: CreateContactDto) => {
     if(foundContact){
         throw new BadRequest('A contact with the same phone number already exist!');
     }
-
     const contact = {
+        userId: userId,
         name: params.name,
         phone: params.phone,
         email: params.email,
@@ -51,12 +51,30 @@ export const getContact = async (id: string) => {
     return mapContact(foundContact);
 }
 
-export const updateContact = async (id: string, params: UpdateContactDto) => {
+export const updateContact = async (userId:string, id: string, params: UpdateContactDto) => {
+    const foundContact = await Contact.findById(id);
+    if(!foundContact){
+        throw new BadRequest('Not found');
+    }
+
+    if(foundContact.userId.toString() !== userId){
+        throw new BadRequest('Not Allowed');
+    }
+
     const updatedContact = await Contact.findByIdAndUpdate(id, params);
     return mapContact(updatedContact);
 }
 
-export const deleteContact = async (id: string) => {
+export const deleteContact = async (userId: string, id: string) => {
+    const foundContact = await Contact.findById(id);
+    if(!foundContact){
+        throw new BadRequest('Not found');
+    }
+
+    if(foundContact.userId.toString() !== userId){
+        throw new BadRequest('Not Allowed');
+    }
+
     const deletedContact = await Contact.findByIdAndDelete(id);
     return mapContact(deletedContact);
 }
